// src/generated/models/Oman_qpcfk_questionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Oman_qpcfk_questionsstatecode = {
  0: 'Actif',
  1: 'Inactif'
} as const;
export type Oman_qpcfk_questionsstatecode = keyof typeof Oman_qpcfk_questionsstatecode;
export const Oman_qpcfk_questionsstatuscode = {
  1: 'Actif',
  2: 'Inactif'
} as const;
export type Oman_qpcfk_questionsstatuscode = keyof typeof Oman_qpcfk_questionsstatuscode;

export interface Oman_qpcfk_questionsBase {
  importsequencenumber?: number;
  oman_fkid: string;
  "oman_LCategory@odata.bind": string;
  "oman_LQuestion@odata.bind": string;
  "oman_LReponse@odata.bind": string;
  "oman_LReponseInfo@odata.bind"?: string;
  "oman_LSection@odata.bind"?: string;
  oman_name: string;
  oman_qpcfk_questionid: string;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Oman_qpcfk_questionsstatecode;
  statuscode?: Oman_qpcfk_questionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Oman_qpcfk_questions extends Oman_qpcfk_questionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  oman_lcategoryname?: string;
  oman_lquestionname?: string;
  oman_lreponseinfoname?: string;
  oman_lreponsename?: string;
  oman_lsectionname?: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  oman_lcategory?: object;
  _oman_lcategory_value?: string;
  oman_lquestion?: object;
  _oman_lquestion_value?: string;
  oman_lreponse?: object;
  _oman_lreponse_value?: string;
  oman_lreponseinfo?: object;
  _oman_lreponseinfo_value?: string;
  oman_lsection?: object;
  _oman_lsection_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}